FIRST all working version FUTHER only features will be added

+++ filesystem.h
@@ -4,7 +4,7 @@
 #include "state.h"
 
 // Функции для работы с файловой системой
-void fs_get_files(const char *path, FileInfo **files, size_t *count);
+void fs_get_files(const char *path, FileInfo **files, long long *count);
 void fs_open_dir(AppState *state, const char *dir_name);
 void fs_open_file(AppState *state, const char *file_name);
